+++ samples/sample/MainScene.h
@@ -33,19 +33,19 @@ namespace Game
         MainScene();
         ~MainScene();
 
-        void resize(const glm::ivec2& newSize) override;
+        void resize(const glm::vec2& newSize) override;
 
         void update(double time) override;
         void draw(Engine::IRenderer* renderer) const override;
 
-        bool onTouchBegan(int fingerIndex, const glm::ivec2& position) override;
-        void onTouchMoved(int fingerIndex, const glm::ivec2& position) override;
+        bool onTouchBegan(int fingerIndex, const glm::vec2& position) override;
+        void onTouchMoved(int fingerIndex, const glm::vec2& position) override;
         void onTouchEnded(int fingerIndex) override;
         void onTouchCancelled(int fingerIndex) override;
 
     private:
         Engine::OrbitCameraPtr mCamera;
         Engine::MeshPtr mMesh;
-        glm::ivec2 mPrevTouchPosition;
+        glm::vec2 mPrevTouchPosition;
     };
 }
